Restructure repository and introduce internal package
