+++ internal/controllers/kubelb/suite_test.go
@@ -24,10 +24,10 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	v1alpha12 "k8c.io/kubelb/pkg/api/kubelb.k8c.io/v1alpha1"
-	"k8c.io/kubelb/pkg/envoy"
-	"k8c.io/kubelb/pkg/kubelb"
-	portlookup "k8c.io/kubelb/pkg/port-lookup"
+	v1alpha12 "k8c.io/kubelb/api/kubelb.k8c.io/v1alpha1"
+	"k8c.io/kubelb/internal/envoy"
+	"k8c.io/kubelb/internal/kubelb"
+	portlookup "k8c.io/kubelb/internal/port-lookup"
 
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
